Fix labels on group node splitters when inside that group
